- Added Reporter.kt documentation - Moved List<Reporter>.report() extension function to utils
