+++ src/jvmMain/kotlin/com/david/lamar/kneat/util/ExtensionFunctions.kt
@@ -1,7 +1,12 @@
 package com.david.lamar.kneat.util
 
+import com.david.lamar.kneat.Pipeline
+import com.david.lamar.kneat.configuration.GenomeConfiguration
+import com.david.lamar.kneat.genome.Genome
 import com.david.lamar.kneat.genome.genes.Gene
+import com.david.lamar.kneat.genome.genes.attributes.Attribute
 import com.david.lamar.kneat.genome.network.NetworkElement
+import com.david.lamar.kneat.util.reporting.Reporter
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
 import kotlinx.coroutines.withContext
@@ -69,14 +74,41 @@ suspend inline fun <T, R> Iterable<T>.mapParallel(crossinline transform: suspend
     list.map { it.await() }
 }
 
-    for (item in this@sortedParallelBy) {
-        newSelector[item] = async { selector(item) }
+/**
+ * Helper function to retrieve an attribute from a map of attributes according to the indexed
+ * name of the attribute being sought out.
+ */
+inline fun <reified T> Map<String, Attribute<out Any, out GenomeConfiguration.AttributeConfiguration<out Any>>>.getAs(key: String) : T {
+    return this.getOrElse(key) { error("Invalid gene configuration. ") } as T
+}
+
+/**
+ * A helper function to report for an entire list of reporters all at once. It's not uncommon for
+ * a developer to want reporters for different things; having to loop through every place we want to call
+ * a reporting method is cumbersome. This simplifies that process
+ */
+fun List<Reporter>.report() : Reporter {
+    val list = this
+
+    return object : Reporter {
+        override fun info(info: String) = list.forEach { it.info(info) }
+        override fun warn(warning: String) = list.forEach { it.warn(warning) }
+        override fun error(error: String, throwable: Throwable?) = list.forEach { it.error(error, throwable) }
+        override fun onCreatingInitialPopulation() = list.forEach { it.onCreatingInitialPopulation() }
+        override fun onSpeciating() = list.forEach { it.onSpeciating() }
+        override fun onSpeciated() = list.forEach { it.onSpeciated()}
+        override fun onEvaluating(genome: Genome) = list.forEach { it.onEvaluating(genome) }
+        override fun onEvaluated() = list.forEach { it.onEvaluated() }
+        override fun onGenerationStarted() = list.forEach { it.onGenerationStarted() }
+        override fun onGenerationFinished() = list.forEach { it.onGenerationFinished()}
+        override fun onCulling() = list.forEach { it.onCulling() }
+        override fun onCulled() = list.forEach { it.onCulled() }
+        override fun onReproducing() = list.forEach { it.onReproducing() }
+        override fun onReproduced() = list.forEach { it.onReproduced()}
+        override fun onExtinction() = list.forEach { it.onExtinction() }
+        override fun onSolutionFound(fittest: Genome?) = list.forEach { it.onSolutionFound(fittest) }
+        override fun onPaused(willResumeState: Pipeline.PipelineState) = list.forEach { it.onPaused(willResumeState) }
+        override fun onSaving(wilLResumeState: Pipeline.PipelineState) = list.forEach { it.onSaving(wilLResumeState) }
+        override fun onLoading() = list.forEach { it.onLoading() }
     }
-
-    val selectors = newSelector.map { (item, wait) ->
-        item to wait.await()
-    }.toMap()
-
-
-    return@withContext sortedBy { selectors[it] }
 }
